002 - Python/002-3 - PROD/P01 - Rotina/MAIN/CARGA DE TABELAS/CLM por FTP [Testes]/requirements.txt
paramiko>=2.11.0
pandas>=1.5.0
pyarrow>=10.0.0
unidecode>=1.3.0
google-cloud-storage>=2.0.0
